user-service/requirements.txt
# user-service/requirements.txt
flask==3.0.0
mysql-connector-python==8.3.0
gunicorn==21.2.0
pytest==7.4.0
pytest-flask==1.3.0